Update v1.2.5
Fixed Auto clicker to allow it to stop and center the clicking animation also made it so that the gui appears over the animation not under. Added the feature to only add items that are real items and if there not it will not add it and tell you to check again to prevent account glitches (N/A account value).

Signed-off-by: pᴉʞ ɥsɐɹʇ ɹnoʎ <91683715+Your-Trash-kid@users.noreply.github.com>